+++ test/testmcdm.jl
@@ -1,6 +1,7 @@
 @testset "MCDM functions" verbose = true begin
 
 include("mcdm/testram.jl")
+include("mcdm/testrim.jl")
 include("mcdm/testbestworst.jl")
 include("mcdm/testidocriw.jl")
 include("mcdm/testcilos.jl")