+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.8.4 (Upcoming release)
 
+- Simplify code and documentation of CRADIS method.
+
 
 ### 0.8.3